NaoDebugger Unresponsive Temporary Fix
The NaoDebugger now remains responsive when launching an event.
In Walk.WalkTo, the proxy is using non-blocking method calls.
Also, no use is made of EventQueue, as that still locks the GUI thread
even while using non-blocking calls of the proxies.
Consider making EventQueue run not in a separate thread, but rather have
an progress() or iterate() method that will deal with the next event in
queue.